requirements.txt
# Core
torch>=2.2,<2.4
numpy==1.26.4
scipy>=1.10
pandas>=2.1
pyyaml>=6.0

# Data handling
xarray>=2023.6
netCDF4>=1.6

# Metrics & utils
scikit-learn==1.4.2
tqdm>=4.66
einops>=0.7

# Visualization
matplotlib>=3.8
seaborn>=0.13
imageio>=2.34

# App
streamlit>=1.36

# Optional (commented to ease Windows install)
# rasterio>=1.3
# rioxarray>=0.15
# opencv-python>=4.9
# cartopy>=0.22
# geopandas>=0.14
